% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{check_structure}
\alias{check_structure}
\title{Helper to check the structure columns define a sound hierarchy}
\usage{
check_structure(df)
}
\arguments{
\item{df}{data frame containing the structure columns image_label, slide_label,
woodpiece_label, tree_label and site_label}
}
\description{
Helper to check the structure columns define a sound hierarchy
}
\keyword{internal}
