% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_data.R
\name{collect_cells_data}
\alias{collect_cells_data}
\title{Collect raw cells output data}
\usage{
collect_cells_data(df_structure)
}
\arguments{
\item{df_structure}{Dataframe containing filenames and data structure.}
}
\value{
A dataframe with the raw data from all files combined (relevant columns only).
}
\description{
Read and combine data from all cell output files
}
