% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_metadata.R
\name{collect_metadata_from_files}
\alias{collect_metadata_from_files}
\title{Read and combine raw metadata}
\usage{
collect_metadata_from_files(df_structure, roxas_version = "classic", tz = NULL)
}
\arguments{
\item{df_structure}{Dataframe with all input filenames}

\item{roxas_version}{ROXAS version (required to correctly extract info from settings files)}

\item{tz}{Timezone for date conversion of settings created at date (default: system tz)}
}
\value{
A dataframe containing the extracted settings data.
}
\description{
Collect the ROXAS settings and image exif data from all raw files and
combine into one dataframe with the data structure.
}
