% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rxs_metadata.R
\name{combine_rxs_metadata}
\alias{combine_rxs_metadata}
\title{Combine raw ROXAS metadata}
\usage{
combine_rxs_metadata(df_structure, df_settings)
}
\arguments{
\item{df_structure}{Dataframe with all input filenames and data structure.}

\item{df_settings}{Dataframe with extracted ROXAS settings data.}
}
\value{
A \code{QWAmetadata} object with the joined image-level metadata in \verb{$images}.
}
\description{
Combine the extracted ROXAS filenames plus data structure (from
\code{\link[=extract_data_structure]{extract_data_structure()}}) and ROXAS settings and image exif data (from
\code{\link[=collect_settings_data]{collect_settings_data()}}) into one raw metadata dataframe.
}
