% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QWAdata.R
\name{complete_measures}
\alias{complete_measures}
\title{Complete QWA measures}
\usage{
complete_measures(QWA_data)
}
\arguments{
\item{QWA_data}{a \code{QWAdata} object containing the cells and rings data frames}
}
\value{
a \code{QWAdata} object with the updated cells and rings dataframes with the new measures
}
\description{
Some additional cell measures are calculated based on the existing data:
\itemize{
\item tca: la + cwa
\item rwd2: cwtrad/drad
\item dcwt
\item raddistr.st: raddistr standardized by mrw
\item cwtall.adj
\item cdrad, cdtan, cdratio
\item sector100
\item ew_lw: indicates if it is an EW or LW cell (based on Mork index <1 for EW)
And for the rings, we add
\item eww and lww estimates  (based on Mork index of 1)
If any of these columns already exist in the data they are NOT recalculated, only
the missing measures are added.
}
}
