% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QWAimages.R
\name{complete_QWAimages}
\alias{complete_QWAimages}
\title{Complete a QWAimages object}
\usage{
complete_QWAimages(x)
}
\arguments{
\item{x}{A \link{QWAimages} object.}
}
\value{
A \link{QWAimages} object extended with all possible columns.
}
\description{
Add any missing columns to a \link{QWAimages} object, including optional columns.
New columns are intitialized as \code{NA}.
}
\seealso{
\code{\link[=complete_QWAmetadata]{complete_QWAmetadata()}}, \code{\link[=QWAimages]{QWAimages()}}
}
