% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{convert_column}
\alias{convert_column}
\title{Helper to convert a column to the target class}
\usage{
convert_column(x, target_class)
}
\arguments{
\item{x}{The column to convert.}

\item{target_class}{The target class to convert to}
}
\description{
Helper to convert a column to the target class
}
\keyword{internal}
