% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rxs_metadata.R
\name{convert_settings_dates}
\alias{convert_settings_dates}
\title{Handle date conversion for ROXAS settings created at dates}
\usage{
convert_settings_dates(
  date_strings,
  orders = c("\%d.\%m.\%Y \%H:\%M:\%S", "\%d/\%m/\%Y \%H:\%M"),
  tz = Sys.timezone(),
  exact = TRUE
)
}
\arguments{
\item{date_strings}{Vector of date strings to be converted.}

\item{orders}{One or more date formats to be used for parsing,
see \code{\link[lubridate:parse_date_time]{lubridate::parse_date_time()}}.}

\item{tz}{Timezone for the converted dates (default: system timezone).}

\item{exact}{Whether to enforce exact matching of the provided format(s), or
allow lubridate some flexibility to guess.}
}
\value{
A vector of POSIXct dates.
}
\description{
The Date created field in the raw ROXAS settings files may have different
formats depending on the ROXAS version and user locale.
}
