% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_profiles.R
\name{create_bands_dt}
\alias{create_bands_dt}
\title{Helper function to create the moving band definitions for given a mrw,
andwidth and stepsize}
\usage{
create_bands_dt(mrw_val, bandwidth, stepsize, band_rebound = TRUE)
}
\arguments{
\item{mrw_val}{numeric, the mrw value to create bands for}

\item{bandwidth}{numeric, the bandwidth of each band}

\item{stepsize}{numeric, the step size between band starts}

\item{band_rebound}{logical, if TRUE, the last band is rebounded to end exactly at mrw_val}
}
\value{
a data.table with start and end positions of each band for the given mrw_val
}
\description{
Helper function to create the moving band definitions for given a mrw,
andwidth and stepsize
}
\keyword{internal}
