% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coverage.R
\name{create_coverage_plots}
\alias{create_coverage_plots}
\title{Create coverage plots for all trees}
\usage{
create_coverage_plots(df_rings, save_plot = TRUE, path_out = "./")
}
\arguments{
\item{df_rings}{rings dataframe}

\item{save_plot}{should the plot be written to disk under path_out}

\item{path_out}{path where the plot should be saved.}
}
\description{
Function to create coverage plots for all unique treecodes in the provided
rings df (usually QWA_data$rings) in a for loop. Plots can be shown and/or
saved to a specified path.
}
