% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.R
\name{flag_duplicate_rings}
\alias{flag_duplicate_rings}
\title{Flag duplicate rings due to overlapping images from the same woodpiece}
\usage{
flag_duplicate_rings(df_rings_log)
}
\arguments{
\item{df_rings_log}{the dataframe containing the rings data}
}
\value{
the input dataframe with additional logical column 'duplicate_ring'
and integer column 'duplicate_sel' with the selected ring among duplicates
}
\description{
This function identifies duplicate rings, i.e. the same annual ring being
captured in two or more images from the same woodpiece due to overlaps.
It also ranks the duplicated rings by the number of cells they contain (in
decreasing order), since for chronologies, we generally want to keep only the
ring with the most cells (rank 1).
}
\keyword{internal}
