% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.R
\name{flag_incomplete_rings}
\alias{flag_incomplete_rings}
\title{Flag incomplete rings at tangential image borders.}
\usage{
flag_incomplete_rings(df_rings_log, df_cells_all, df_meta)
}
\arguments{
\item{df_rings_log}{the dataframe containing the rings data}

\item{df_cells_all}{the dataframe containing all cells data}

\item{df_meta}{the dataframe containing the metadata of the images (needed
for the spatial resolution)}
}
\value{
the input dataframe with additional logical columns 'incomplete_ring'
}
\description{
This function determines which of the inner- and outermost rings of each
image are incomplete (i.e. extend across the image boundaries).
For the innermost rings, we rely on the position and shape of the estimated
inner ring border (see helper function \code{check_incomplete_innermost}). For the
outermost rings, we check the MRW value, since ROXAS only estimates an MRW if
it can detect an outer ring boundary.
}
