% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{get_tbl_props}
\alias{get_tbl_props}
\title{Helper to extract specific column definitions from a imported schema}
\usage{
get_tbl_props(tbl_schema)
}
\description{
Helper to extract specific column definitions from a imported schema
}
\keyword{internal}
