% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coverage.R
\name{plot_woodpiece_coverage}
\alias{plot_woodpiece_coverage}
\title{Plot overview of annual coverage of images}
\usage{
plot_woodpiece_coverage(woodpiece, df_rings)
}
\arguments{
\item{woodpiece}{the woodpiece (core) for which the plot should be created}

\item{df_rings}{the dataframe containing the rings data (including flags)}
}
\description{
For a single tree, each image covers a span of years. This function allows
to visualize which annual rings are covered by which image and which rings
have been flagged due to which issues.
}
