% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rxs_data.R
\name{read_cells_output}
\alias{read_cells_output}
\title{Read single cell output file}
\usage{
read_cells_output(file_cells, selcols_cells, colname_variants)
}
\arguments{
\item{file_cells}{filename to be read}

\item{selcols_cells}{character vector of the required columns to keep}

\item{colname_variants}{named character vector of any variant column names to rename}
}
\value{
A dataframe with the raw data (relevant columns only).
}
\description{
Helper function to read a single (classical) ROXAS cells output file
}
\keyword{internal}
