% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rxs_data.R
\name{read_output_file}
\alias{read_output_file}
\title{Read single cell or rings measurements file}
\usage{
read_output_file(filename, selcols, colname_variants, delim)
}
\arguments{
\item{filename}{filename to be read}

\item{selcols}{character vector of the required columns to keep}

\item{colname_variants}{named character vector of any variant column names to rename}

\item{delim}{File delimiter (usually \code{"\\t"} for ROXAS, \code{";"} for ROXAS AI).}
}
\value{
A dataframe with the raw data (relevant columns only, cleaned names).
}
\description{
Helper function to read a single ROXAS (AI) output (table) file
}
\keyword{internal}
