% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwa_IO.R
\name{read_QWAmetadata}
\alias{read_QWAmetadata}
\title{Read a QWAmetadata object from a JSON file}
\usage{
read_QWAmetadata(
  file,
  roxas_version = NULL,
  allow_missing_req = FALSE,
  add_missing_opt = FALSE
)
}
\arguments{
\item{file}{Path to a \code{.json} or \code{.json.gz} file of QWAmetadata.}

\item{roxas_version}{ROXAS software version (optional). Should usually be inferred
from the data directly, but can be overwritten here (\code{"roxas"} or `"roxas_ai"``).}

\item{allow_missing_req}{If \code{FALSE} (default), missing required columns raise an error;
if \code{TRUE}, they raise a warning and are added as \code{NA}.}

\item{add_missing_opt}{If \code{FALSE} (default), missing optional columns raise a warning only;
if \code{TRUE} they raise a warning and are added as \code{NA}.}
}
\value{
A \link{QWAmetadata} object.
}
\description{
Read a QWAmetadata object from a JSON file
}
\seealso{
\code{\link[=write_QWAmetadata]{write_QWAmetadata()}}
}
