% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rxs_data.R
\name{read_rings_output}
\alias{read_rings_output}
\title{Read single rings output file}
\usage{
read_rings_output(file_rings, selcols_rings, colname_variants)
}
\arguments{
\item{file_rings}{filename to be read}

\item{colname_variants}{named character vector of any variant column names to rename}

\item{selcols_cells}{character vector of the required columns to keep}
}
\value{
A dataframe with the raw data (relevant columns only).
}
\description{
Helper function to read a single (classical) ROXAS rings output file
}
\keyword{internal}
