% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{resolve_schema}
\alias{resolve_schema}
\title{Read and resolve JSON Schema from file}
\usage{
resolve_schema(schema_path)
}
\arguments{
\item{schema_path}{Path to the JSON schema file}
}
\value{
A list containing the fully resolved schema
}
\description{
Read and resolve JSON Schema from file
}
