% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_metadata.R
\name{subset_woodpiece_codes}
\alias{subset_woodpiece_codes}
\title{Include or exclude woodpiece codes from data collection}
\usage{
subset_woodpiece_codes(
  df_structure,
  include_codes = NULL,
  exclude_codes = NULL
)
}
\arguments{
\item{df_structure}{The dataframe containing all input filenames and data structure.}

\item{include_codes}{EITHER provide a vector of the woodpiece codes to be included}

\item{exclude_codes}{OR provide a vector of the woodpiece codes to be excluded}
}
\value{
filtered df_structure.
}
\description{
This function filters the data structure dataframe based on the woodpiece codes provided
}
