% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{validate_schema}
\alias{validate_schema}
\title{Validate against json schema}
\usage{
validate_schema(df, schema = c("images", "authors"), warn_only = TRUE)
}
\arguments{
\item{df}{the data frame to validate.}

\item{schema}{the name of the QWAmetadata element under question}

\item{warn_only}{abort on errors}
}
\value{
passes if validation ok
}
\description{
Validate against json schema
}
