% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QWAmetadata.R
\name{write_QWAmetadata}
\alias{write_QWAmetadata}
\title{Write a QWAmetadata object to a JSON file}
\usage{
write_QWAmetadata(x, file, compress = FALSE, overwrite = TRUE)
}
\arguments{
\item{x}{A \link{QWAmetadata} object.}

\item{file}{Path to the output file.}

\item{compress}{If \code{TRUE}, write compressed \code{.gz} file. (default \code{FALSE}).}

\item{overwrite}{Allow to overwrite existing files? (default \code{TRUE}).}
}
\value{
The output file path, invisibly.
}
\description{
Write a QWAmetadata object to a JSON file
}
\seealso{
\code{\link[=read_QWAmetadata]{read_QWAmetadata()}}
}
