% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwa_IO.R
\name{write_QWAprofile}
\alias{write_QWAprofile}
\title{Write a QWAprofile object to a CSV file}
\usage{
write_QWAprofile(x, file, compress = FALSE, overwrite = TRUE)
}
\arguments{
\item{x}{A \link{QWAprofile} object.}

\item{file}{Path to the output \code{.csv} or \code{.csv.gz} file.}

\item{compress}{If \code{TRUE}, write a compressed \code{.csv.gz} file (default \code{FALSE}).}

\item{overwrite}{Allow overwriting existing files? (default \code{TRUE}).}
}
\value{
The output file path, invisibly.
}
\description{
Write a QWAprofile object to a CSV file
}
\seealso{
\code{\link[=read_QWAprofile]{read_QWAprofile()}}, \code{\link[=QWAprofile]{QWAprofile()}}
}
