% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_resources.R
\name{add_resources}
\alias{add_resources}
\title{Add a resources table to a QWAmetadata object}
\usage{
add_resources(x, path, recursive = FALSE, include_unmatched = FALSE)
}
\arguments{
\item{x}{A \link{QWAmetadata} object.}

\item{resources}{A data frame of resources, typically from \code{\link[=collect_resources]{collect_resources()}}.}
}
\value{
The \link{QWAmetadata} object with the \verb{$resources} component updated.
}
\description{
Stores a resources data frame (typically created by \code{\link[=collect_resources]{collect_resources()}})
as the \verb{$resources} component of a \link{QWAmetadata} object. Any previously
stored resources are replaced.
}
\seealso{
\code{\link[=collect_resources]{collect_resources()}}, \code{\link[=QWAmetadata]{QWAmetadata()}}
}
