% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_profiles.R
\name{calculate_band_profiles}
\alias{calculate_band_profiles}
\title{Calculate band-wise cell parameter profiles}
\usage{
calculate_band_profiles(
  QWA_data,
  bandwidth,
  stepsize,
  sel_cell_params,
  quant_probs = NULL,
  band_rebound = TRUE
)
}
\arguments{
\item{QWA_data}{list containing the rings and cells dataframes}

\item{bandwidth}{the width of the band in microns}

\item{stepsize}{the step size with which the band is shifted, in microns}

\item{sel_cell_params}{character vector with the names of the cell parameters
to include in the profiles (e.g. c("la", "cwttan"))}

\item{quant_probs}{numeric vector with the quantile probabilities to
calculate (e.g. c(0.25, 0.5, 0.75)). If NULL or empty, no quantiles are
calculated (i.e. only the mean).}
}
\value{
a dataframe with the sector-wise profiles of the selected cell
parameters
}
\description{
This function calculates band-wise profiles of selected cell parameters
(means and quantiles) from the cells data. A band is a moving window of
given bandwidth that moves from the start to the end of the ring at the given
stepsize.
}
