% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rxs_metadata.R
\name{collect_image_info}
\alias{collect_image_info}
\title{Read and combine image exif data}
\usage{
collect_image_info(files_images)
}
\arguments{
\item{files_images}{Vector/list of image filenames}
}
\value{
A dataframe containing the extracted data.
}
\description{
Collect the exif data from all image files
}
