% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rxs_data.R
\name{collect_raw_data}
\alias{collect_raw_data}
\title{Collect raw cells and rings output data}
\usage{
collect_raw_data(df_structure)
}
\arguments{
\item{df_structure}{Dataframe containing filenames and data structure.}
}
\value{
A \code{QWAdata} object containing the combined raw data for cells and
rings in \verb{$cells} and \verb{$rings}, respectively.
}
\description{
Read and combine data from all cells and ring output files
}
