% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rxs_data.R
\name{collect_raw_outputs}
\alias{collect_raw_outputs}
\title{Collect raw output data from multiple files}
\usage{
collect_raw_outputs(df_structure, roxas_version, ftype)
}
\arguments{
\item{df_structure}{Data frame of image identifiers (\verb{$image_label}) and
corresponding data file names (\verb{$fname_cells} or \verb{$fname_rings}).}

\item{roxas_version}{The software used to create the files (\code{"roxas"} or \code{"roxas_ai"}).}

\item{ftype}{which type of files should be read (\code{"cells"} or \code{"rings"}).}
}
\value{
A data frame with the relevant raw measurements data from all
files combined.
}
\description{
Read and combine data from multiple cells/rings data files (i.e. the
ROXAS output .txt or ROXAS AI table .csv files with the raw measurements).
Used by \code{\link[=collect_raw_data]{collect_raw_data()}} to read each \code{QWAdata} component.
}
\seealso{
\code{\link[=collect_raw_data]{collect_raw_data()}}
}
