% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rxs_metadata.R
\name{collect_settings_data}
\alias{collect_settings_data}
\title{Read and combine ROXAS settings data}
\usage{
collect_settings_data(files_settings, roxas_version = "classic")
}
\arguments{
\item{files_settings}{Vector/list of ROXAS settings filenames}

\item{roxas_version}{The version of ROXAS used to create the files (classic, AI)}
}
\value{
A dataframe containing the extracted data.
}
\description{
Collect the settings data from all ROXAS settings files
}
