% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.R
\name{complete_cell_measures}
\alias{complete_cell_measures}
\title{Complete QWA measures}
\usage{
complete_cell_measures(QWA_data)
}
\arguments{
\item{QWA_data}{a \code{QWAdata} object containing the cells and rings dataframes}
}
\value{
a \code{QWAdata} object with the updated cells and rings dataframes with the new measures
}
\description{
Some additional cell measures are calculated based on the existing data:
\itemize{
\item tca: la + cwa
\item rwd2: cwtrad/drad
\item dcwt
\item raddistr.st: raddistr standardized by mrw
\item cwtall.adj
\item cdrad, cdtan, cdratio
\item sector100
\item ew_lw: indicates if it is an EW or LW cell (based on Mork index <1)
And for the rings, we add
\item eww and lww estimates  (based on Mork index <1)
}
}
