% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{create_empty_df}
\alias{create_empty_df}
\title{Create an empty dataframe conforming to a JSON schema}
\usage{
create_empty_df(tbl_props, nrows = 0)
}
\arguments{
\item{tbl_props}{The properties from table schema (i.e. after
\code{\link[=resolve_refs]{resolve_refs()}} and \code{\link[=get_tbl_props]{get_tbl_props()}}), with \verb{$properties}
containing the column definitions.}

\item{nrows}{Number of rows to create. If \verb{> 0}, character columns are
filled with \code{""} and logical columns with \code{FALSE}.}
}
\value{
A tibble with columns and R types matching the schema, with
\code{nrows} rows.
}
\description{
Create an empty dataframe conforming to a JSON schema
}
\keyword{internal}
