% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_metadata.R
\name{extract_roxas_settings}
\alias{extract_roxas_settings}
\title{Extract data from a ROXAS settings file}
\usage{
extract_roxas_settings(file_settings, roxas_version = "classic")
}
\arguments{
\item{file_settings}{The file to be read.}

\item{roxas_version}{The version of ROXAS used to create the file (classic, AI)}
}
\value{
A dataframe containing the extracted data.
}
\description{
Helper function to read and extract the relevant metadata from a single ROXAS
settings file
}
