% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rxs_metadata.R
\name{get_roxas_files}
\alias{get_roxas_files}
\title{Identify the ROXAS (AI) files in a directory}
\usage{
get_roxas_files(path_in, roxas_version, exclude_dirs = NULL)
}
\arguments{
\item{path_in}{path of the input directory.}

\item{roxas_version}{which ROXAS version was used to create the files (classic
\code{"roxas"} or new AI version \code{"roxas_ai"}).}

\item{exclude_dirs}{directory names / keywords that should be ignored
when searching for ROXAS files (e.g., "unused_files", optional)}
}
\value{
A data frame with columns \code{prefix}, \code{fname_image}, \code{fname_settings},
\code{fname_cells} and \code{fname_rings} containing the full file paths.
}
\description{
Returns a data frame with the full path names of the relevant ROXAS (AI) files
(i.e., images, cell measurements, ring measurements, and ROXAS settings resp.
ROXAS AI metadata files) found within the given directory.
The function recurses into any subdirectories, except those explicitly excluded.
The files are identified by matching on the standard naming suffixes produced
by ROXAS resp. ROXAS AI.
}
\details{
For ROXAS, the required files per image are:
\itemize{
\item Image file (\verb{\{IMAGEID\}.jpg} or \code{.jpeg})
\item Cells output (\verb{\{IMAGEID\}_Output_Cells.txt})
\item Rings output (\verb{\{IMAGEID\}_Output_Rings.txt})
\item Settings file (\verb{\{IMAGEID\}_ROXAS_Settings.txt})
}

For ROXAS AI, the required files per image are:
\itemize{
\item Image file (\verb{\{IMAGEID\}.scan.jpg} or \code{.jpeg})
\item Cells table (\verb{\{IMAGEID\}.cells_table.csv})
\item Rings table (\verb{\{IMAGEID\}.rings_table.csv})
\item Metadata file (\verb{\{IMAGEID\}.metadata.json})
}
}
