% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roxas_files.R
\name{get_roxas_files}
\alias{get_roxas_files}
\title{Identify the ROXAS (AI) files in a directory}
\usage{
get_roxas_files(path_in, roxas_version, exclude_dirs = NULL)
}
\arguments{
\item{path_in}{The path of the input directory.}

\item{roxas_version}{Which software was used to create the files, either
\code{"roxas"} or \code{"roxas_ai"}.}

\item{exclude_dirs}{Optional character vector of subdirectory names that
should not be searched for ROXAS (AI) files (e.g., \code{"unused_files"}).}
}
\value{
A data frame with columns \code{prefix}, \code{fname_image}, \code{fname_settings},
\code{fname_cells} and \code{fname_rings} containing the common prefixes and full
file paths.
}
\description{
Returns a data frame with the full paths of the four types of relevant
ROXAS (AI) files (i.e., images, cell measurements, ring measurements, and
ROXAS settings resp. ROXAS AI metadata files) found within the given
directory.
Recurses into any subdirectories except those explicitly excluded.
The files are identified by matching on the standard naming suffixes produced
by ROXAS resp. ROXAS AI, see the Details below.
}
\details{
For ROXAS, the required files per image are:
\itemize{
\item The original image file (\verb{\{IMAGEID\}.jpg} or \code{.jpeg}).
\item The ROXAS settings file (\verb{\{IMAGEID\}_ROXAS_Settings.txt}).
\item The cell measurements output (\verb{\{IMAGEID\}_Output_Cells.txt}).
\item The ring measurements output (\verb{\{IMAGEID\}_Output_Rings.txt}).
}

For ROXAS AI, the required files per image are:
\itemize{
\item The original image file (\verb{\{IMAGEID\}.scan.jpg} or \code{.jpeg}).
\item The ROXAS AI metadata file (\verb{\{IMAGEID\}.metadata.json}).
\item The cell measurements table (\verb{\{IMAGEID\}.cells_table.csv}).
\item The ring measurements table (\verb{\{IMAGEID\}.rings_table.csv}).
}
}
