% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rxs_metadata.R
\name{get_roxas_files}
\alias{get_roxas_files}
\title{Identify ROXAS data files}
\usage{
get_roxas_files(path_in, roxas_version = c("roxas", "roxas_ai"))
}
\arguments{
\item{path_in}{path of the input directory.}
}
\value{
A list of lists containing filepaths of images and ROXAS output
files (cells, rings, settings).
}
\description{
Given a path to a directory, returns a nested list with the full filepaths of
all required ROXAS files (images, cell outputs, ring outputs, settings) found
within (including subdirectories).
}
