% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_metadata.R
\name{get_roxas_files}
\alias{get_roxas_files}
\title{Get list of input files}
\usage{
get_roxas_files(path_in)
}
\arguments{
\item{path_in}{path of the input directory.}
}
\value{
A list of lists containing filepaths of images and ROXAS output
files (cells, rings, settings).
}
\description{
Given a path_in, get the paths and filenames of the ROXAS files to be
processed.
}
