% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rxs_metadata.R
\name{get_structure_from_filenames}
\alias{get_structure_from_filenames}
\alias{extract_data_structure}
\title{Extract the data structure from file names}
\usage{
get_structure_from_filenames(
  filenames,
  pattern =
    "(?<site>[[:alnum:]]+)_(?<species>[[:alnum:]]+)_(?<tree>[[:alnum:]][[:alnum:]])(?<woodpiece>[[:alnum:]]*)_(?<slide>[[:alnum:]]+)_(?<image>[[:alnum:]]+)",
  site_label = NULL,
  species_code = NULL
)

extract_data_structure(
  files,
  pattern =
    "(?<site>[[:alnum:]]+)_(?<species>[[:alnum:]]+)_(?<tree>[[:alnum:].]+)_(?<slide>[[:alnum:]]+)_(?<image>[[:alnum:]]+)",
  site_label = NULL,
  species_code = NULL
)
}
\arguments{
\item{filenames}{The vector of file names.}

\item{pattern}{The labeling pattern followed by the file names, a regex
with named groups.}

\item{site_label}{Optional site code to be used if it is not part of the
pattern, only if all images are from the same site.}

\item{species_code}{Optional species code to be used if it is not part of the
pattern, only if all images are from the same species.}

\item{files}{The named list of file vectors returned by \code{\link[=get_roxas_files]{get_roxas_files()}}.}
}
\value{
A dataframe containing the extracted data structure (plus the
original file names for \code{\link[=extract_data_structure]{extract_data_structure()}}).
}
\description{
Extract the hierarchical structure of the data (i.e., which images belong to
which slide, woodpiece, tree, site) from the given file names based on a
provided regex labeling pattern.
The extracted site, species, tree, woodpiece, slide, image identifiers are
joined by underscores to derive unique codes for each level of the data
structure for the subsequent calculations
(e.g. \verb{tree_code = \{site\}_\{species\}_\{tree\}}).

The default pattern assumes that all file names follow the labeling scheme
\verb{\{site\}_\{species\}_\{tree\}\{woodpiece\}_\{slide\}_\{image\}},
where \code{{tree}} is a two-character code, and \code{{woodpiece}} is optional, as
suggested in \href{https://doi.org/10.3389/fpls.2016.00781}{Fonti et al. (2025)}.

However, a different labeling pattern may be used, provided that it allows to
uniquely identify each image in the list of file names and is consistently
applied across all files.
Not all hierarchical components are mandatory in the pattern, as long as the
images can be uniquely identified (e.g. if exactly one woodpiece was
extracted per tree, there might not be a specific woodpiece identifier,
or if all data is from the same site, there might not be a site identifier).
If all images pertain to the same site and/or the same species, and these
components are not part of the labeling pattern, they may also be provided
via the \code{site_label} and \code{species_code} input parameters.
Here, \code{species_code} should follow the standard ITRDB species codes.

\code{\link[=extract_data_structure]{extract_data_structure()}} is a wrapper around \code{\link[=get_structure_from_filenames]{get_structure_from_filenames()}}
that uses the list of ROXAS (AI) files from \code{\link[=get_roxas_files]{get_roxas_files()}} as input,
applies the hierarchy extraction to the \verb{$fname_image} component,
and appends the original filepaths as columns to the structure data frame output.
}
\examples{
# with the default pattern ({site}_{species}_{tree}{woodpiece}_{slide}_{image})
filenames <- c(
  "SITEA_PISY_01A_1_1.jpg",
  "SITEA_PISY_01A_1_2.jpg",
  "SITEA_PISY_01A_2_1.jpg",
  "SITEB_LASI_02A_1_1.jpg",
  "SITEB_LASI_02B_1_1.jpg",
  "SITEB_LASI_03_1_1.jpg"
)
get_structure_from_filenames(filenames)

# custom pattern of the form {tree}_{slide}_{image}
filenames <- c(
 "tree1_sl1_img1.jpg",
 "tree1_sl2_img1.jpg",
 "tree2_sl1_img1.jpg",
 "tree2_sl1_img2.jpg"
 )
 pattern <- "(?<tree>[[:alnum:]].+)_(?<slide>[[:alnum:]]+)_(?<image>[[:alnum:]]+)"
 get_structure_from_filenames(
   filenames, pattern,
   site_label = "SITEA", species_code = "PISY")

}
