% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_structure_from_filenames.R
\name{get_structure_from_filenames}
\alias{get_structure_from_filenames}
\alias{extract_data_structure}
\title{Extract the hierarchical data structure from ROXAS file names}
\usage{
get_structure_from_filenames(
  filenames,
  pattern,
  site_label = NULL,
  species_code = NULL
)

extract_data_structure(files, pattern, site_label = NULL, species_code = NULL)
}
\arguments{
\item{filenames}{The character vector of file names or prefixes.}

\item{pattern}{A regex with named groups (site, species, tree, slide and/or
image) to describe the labeling pattern of the \code{filenames}.}

\item{site_label}{Optional site label to be used if it is not part of the
\code{pattern}, only if all images are from the same site.}

\item{species_code}{Optional species code to be used if it is not part of
the \code{pattern}, only if all images are from the same species.}

\item{files}{The data frame of prefixes and file paths returned by \code{\link[=get_roxas_files]{get_roxas_files()}}.}
}
\value{
A data frame containing the extracted data structure (with the
original file name columns appended for \code{\link[=extract_data_structure]{extract_data_structure()}}).
}
\description{
Extract the hierarchical structure of the data (i.e., which images belong to
which slide, woodpiece, tree, site) from the given file names based on a
provided regex labeling pattern.
The extracted site, species, tree, woodpiece, slide, image identifiers are
joined by underscores to derive unique codes for each level of the data
structure for the subsequent calculations
(e.g. \verb{tree_code = \{site\}_\{species\}_\{tree\}}).

Any custom labeling pattern may be used, provided that it allows to
uniquely identify each image and is consistently applied across all files.
The first example below assumes the standard labeling pattern
\verb{\{site\}_\{species\}_\{tree\}_\{slide\}_\{image\}},
proposed by \href{https://doi.org/10.3389/fpls.2025.1505389}{Fonti et al. (2025)}.

Not all hierarchical components are mandatory in the pattern, as long as the
images can be uniquely identified (e.g. if exactly one woodpiece was
extracted per tree, tree and woodpiece identifiers might be identical---in
that case, capture the highest possible hierarchical level in the pattern).
If all images pertain to the same site and/or the same species, the site label
or species code can alternatively be provided as input parameters.
A site identifier (via pattern or site_label) is always required, while the
species code is optional. It is recommended to use the ITRDB species codes.

\code{\link[=extract_data_structure]{extract_data_structure()}} is a wrapper around
\code{\link[=get_structure_from_filenames]{get_structure_from_filenames()}} that uses the data frame of ROXAS (AI)
files from \code{\link[=get_roxas_files]{get_roxas_files()}} as input, applies the hierarchy extraction
to the column \code{prefix}, and appends the original file paths as columns to
the returned structure data frame.
}
\examples{
# the common pattern {site}_{species}_{tree}_{slide}_{image}
pattern <- "(?<site>[[:alnum:]]+)_(?<species>[[:alnum:]]+)_(?<tree>[[:alnum:]]+)_(?<slide>[[:alnum:]]+)_(?<image>[[:alnum:]]+)"
filenames <- c(
  "path/to/file/SITEA_PISY_01_1_1.jpg",
  "path/to/file/SITEA_PISY_01_1_2.jpg",
  "path/to/file/SITEA_PISY_01_2_1.jpg",
  "path/to/file/SITEB_LASI_02_1_1.jpg",
  "path/to/file/SITEB_LASI_03_1_1.jpg"
)
get_structure_from_filenames(filenames, pattern)

# pattern of the form {tree}-{slide}_{image}
# NOTE: filenames without full paths or extension also work
filenames <- c(
 "tree1-sl1_img1",
 "tree1-sl2_img1",
 "tree2-sl1_img1",
 "tree2-sl1_img2"
 )
 pattern <- "(?<tree>[[:alnum:]].+)-(?<slide>[[:alnum:]]+)_(?<image>[[:alnum:]]+)"
 get_structure_from_filenames(
   filenames, pattern,
   site_label = "SITEA", species_code = "LASI")

}
