% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_shiny_apps.R
\name{launch_metadata_app}
\alias{launch_metadata_app}
\title{Launch the metadata Shiny app}
\usage{
launch_metadata_app(server_run = FALSE)
}
\arguments{
\item{server_run}{Logical, if TRUE, runs the app in server mode (for hosting on shiny server).}
}
\description{
Launch the metadata Shiny app
}
