% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_data.R
\name{read_cells_output}
\alias{read_cells_output}
\title{Read single cell output file}
\usage{
read_cells_output(file_cells)
}
\arguments{
\item{file_cells}{filename to be read}
}
\value{
A dataframe with the raw data (relevant columns only).
}
\description{
Helper function
}
