% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_QWAdata}
\alias{read_QWAdata}
\title{Read already processed QWA data from CSV files}
\usage{
read_QWAdata(
  file_path,
  dataset_name = NULL,
  file_rings = NULL,
  file_cells = NULL
)
}
\description{
Read already processed QWA data from CSV files
}
