% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QWAmetadata.R
\name{read_QWAmetadata}
\alias{read_QWAmetadata}
\title{Read a QWAmetadata object from a JSON file}
\usage{
read_QWAmetadata(file)
}
\arguments{
\item{file}{Path to a \code{.json} or \code{.json.gz} file of QWAmetadata.}
}
\value{
A \link{QWAmetadata} object.
}
\description{
Read a QWAmetadata object from a JSON file
}
\seealso{
\code{\link[=write_QWAmetadata]{write_QWAmetadata()}}
}
