% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_data.R
\name{read_rings_output}
\alias{read_rings_output}
\title{Read single rings output file}
\usage{
read_rings_output(file_rings)
}
\arguments{
\item{file_rings}{filename to be read}
}
\value{
A dataframe with the raw data (relevant columns only).
}
\description{
Helper function
}
