% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rxs_data.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers in cells and rings data}
\usage{
remove_outliers(QWA_data, mute_info = FALSE)
}
\arguments{
\item{QWA_data}{a \code{QWAdata} object containing the cells and rings dataframes}
}
\value{
A \code{QWAdata} object with the 'negative' outliers/error codes replaced by NAs
}
\description{
ROXAS does some automatic outlier detection and replaces the found outliers
in cell and ring measures with negative values. This function replaces
these 'negatives' with NA.
ROXAS also uses error codes (-9999, -999, -99) to indicate aborted analysis,
which in newer software versions are alredy set to NA in the .txt output files.
In any case, the removal of negative values also takes care of these.
}
