% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers in cells and rings data}
\usage{
remove_outliers(QWA_data)
}
\arguments{
\item{QWA_data}{a list containing the cells and rings dataframes}
}
\value{
QWA_data with the 'negative' outliers replaced by NAs
}
\description{
ROXAS does some automatic outlier detection and replaces the found outliers
in cell (and ring?) measures by their negative value. This function replaces
these 'negatives' with NA.
}
