% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxs2xcell-package.R
\docType{package}
\name{rxs2tria-package}
\alias{rxs2tria}
\alias{rxs2tria-package}
\title{rxs2tria: Preprocess ROXAS Data for XCELL}
\description{
This package contains the functionalities to preprocess and prepare the ROXAS output files for upload to the XCELL database.
}
\author{
\strong{Maintainer}: Mara Naegelin \email{mara.naegelin@wsl.ch} (\href{https://orcid.org/0000-0002-0490-7510}{ORCID})

}
\keyword{internal}
