% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwa_IO.R
\name{write_QWAimages}
\alias{write_QWAimages}
\title{Write a QWAimages object to a CSV file}
\usage{
write_QWAimages(x, file, compress = FALSE, overwrite = TRUE)
}
\arguments{
\item{x}{A \link{QWAimages} object.}

\item{file}{Path to the output \code{.csv} or \code{.csv.gz} file.}

\item{compress}{If \code{TRUE}, write compressed \code{.csv.gz} file. (default \code{FALSE})}

\item{overwrite}{Allow to overwrite existing files? (default \code{TRUE}).}
}
\value{
The output file path, invisibly.
}
\description{
Write a QWAimages object to a CSV file
}
\seealso{
\code{\link[=read_QWAimages]{read_QWAimages()}}, \link{QWAimages}
}
