% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QWAimages.R
\name{write_QWAimages}
\alias{write_QWAimages}
\title{Write a QWAimages object to a CSV file.}
\usage{
write_QWAimages(x, file, compress = FALSE, overwrite = TRUE)
}
\arguments{
\item{x}{A \link{QWAimages} object.}

\item{file}{Path to the output file.}

\item{compress}{If \code{TRUE}, write compressed \code{.gz} file (default \code{FALSE}).
Note that if you add an compression-related extension to the file name,
vroom will automatically compress the output regardless of this parameter.}

\item{overwrite}{Allow to overwrite existing files? (default \code{TRUE}).}
}
\value{
The output file path, invisibly.
}
\description{
Write a QWAimages object to a CSV file.
}
\seealso{
\code{\link[=read_QWAimages]{read_QWAimages()}}, \link{QWAimages}
}
